@@ -1,28 +1,31 @@
 module github.com/krateoplatformops/rest-dynamic-controller
 
-go 1.23.3
-
-toolchain go1.23.4
+go 1.24.0
 
 require (
 	github.com/gobuffalo/flect v1.0.3
 	github.com/krateoplatformops/snowplow v0.0.0-20250311104630-6e215130151f
-	github.com/krateoplatformops/unstructured-runtime v0.1.1
-	github.com/lucasepe/httplib v0.2.2
-	github.com/pb33f/libopenapi v0.16.8
-	github.com/rs/zerolog v1.33.0
-	k8s.io/api v0.32.1
-	k8s.io/apimachinery v0.32.1
-	k8s.io/client-go v0.32.1
+	github.com/krateoplatformops/unstructured-runtime v0.1.2
+	github.com/pb33f/libopenapi v0.21.8
+	github.com/rs/zerolog v1.34.0
+	github.com/stretchr/testify v1.10.0
+	k8s.io/api v0.32.3
+	k8s.io/apimachinery v0.32.3
+	k8s.io/client-go v0.32.3
 	sigs.k8s.io/controller-runtime v0.20.0
+	sigs.k8s.io/e2e-framework v0.6.0
 )
 
 require (
 	github.com/bahlo/generic-list-go v0.2.0 // indirect
+	github.com/beorn7/perks v1.0.1 // indirect
+	github.com/blang/semver/v4 v4.0.0 // indirect
 	github.com/buger/jsonparser v1.1.1 // indirect
+	github.com/cespare/xxhash/v2 v2.3.0 // indirect
 	github.com/davecgh/go-spew v1.1.2-0.20180830191138-d8f796af33cc // indirect
-	github.com/dprotaso/go-yit v0.0.0-20220510233725-9ba8df137936 // indirect
 	github.com/emicklei/go-restful/v3 v3.12.1 // indirect
+	github.com/evanphx/json-patch/v5 v5.9.0 // indirect
+	github.com/fatih/color v1.18.0 // indirect
 	github.com/fxamacker/cbor/v2 v2.7.0 // indirect
 	github.com/go-logr/logr v1.4.2 // indirect
 	github.com/go-logr/zapr v1.3.0 // indirect
