@@ -1,270 +1,130 @@
 package restclient
 
 import (
+	"bytes"
 	"context"
+	"encoding/json"
+	"fmt"
+	"io"
 	"net/http"
-	"reflect"
+	"net/http/httputil"
+	"os"
 	"strings"
 
-	"fmt"
-
-	unstructuredtools "github.com/krateoplatformops/unstructured-runtime/pkg/tools/unstructured"
-	"github.com/lucasepe/httplib"
-	"github.com/pb33f/libopenapi"
-	v3 "github.com/pb33f/libopenapi/datamodel/high/v3"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 )
 
-type UnstructuredClient struct {
-	IdentifierFields []string
-	SpecFields       *unstructured.Unstructured
-	Server           string
-	DocScheme        *libopenapi.DocumentModel[v3.Document]
-	Auth             httplib.AuthMethod
-	Verbose          bool
-}
-
-// 'field' could be in the format of 'spec.field1.field2'
-func (u *UnstructuredClient) isInSpecFields(field, value string) (bool, error) {
-	fields := strings.Split(field, ".")
-	specs, err := unstructuredtools.GetFieldsFromUnstructured(u.SpecFields, "spec")
-	if err != nil {
-		return false, fmt.Errorf("error getting fields from unstructured: %w", err)
-	}
-
-	val, ok, err := unstructured.NestedFieldCopy(specs, fields...)
-	if err != nil {
-		return false, fmt.Errorf("error getting nested field: %w", err)
-	}
-	if !ok {
-		return false, nil
-	}
-	if reflect.DeepEqual(val, value) {
-		return true, nil
-	}
-
-	return false, nil
-}
-
-func (u *UnstructuredClient) isIdentifier(field string) bool {
-	for _, v := range u.IdentifierFields {
-		if v == field {
-			return true
-		}
-	}
-	return false
-}
-
-type APIError struct {
-	Message   string `json:"message"`
-	TypeKey   string `json:"typeKey"`
-	ErrorCode int    `json:"errorCode"`
-	EventID   int    `json:"eventId"`
-}
-
-type RequestConfiguration struct {
-	Parameters map[string]string
-	Query      map[string]string
-	Body       interface{}
-}
-
-func (u *UnstructuredClient) Get(ctx context.Context, cli *http.Client, path string, opts *RequestConfiguration) (*map[string]interface{}, error) {
-	uri := buildPath(u.Server, path, opts.Parameters, opts.Query)
-	pathItem, ok := u.DocScheme.Model.Paths.PathItems.Get(path)
-	if !ok {
-		return nil, fmt.Errorf("path not found - Get: %s", path)
-	}
-	if len(pathItem.Get.Servers) > 0 {
-		uri = buildPath(pathItem.Get.Servers[0].URL, path, opts.Parameters, opts.Query)
-	}
-
-	err := u.ValidateRequest("GET", path, opts.Parameters, opts.Query)
-	if err != nil {
-		return nil, err
-	}
-	req, err := httplib.Get(uri.String())
-	if err != nil {
-		return nil, err
-	}
-
-	var val map[string]interface{}
-	apiErr := &APIError{}
-
-	httpMethod := "GET"
-	getDoc, ok := pathItem.GetOperations().Get(strings.ToLower(httpMethod))
-	if !ok {
-		return nil, fmt.Errorf("operation not found: %s", httpMethod)
-	}
-
-	validStatusCodes, err := getValidResponseCode(getDoc.Responses.Codes)
-	if err != nil {
-		return nil, err
-	}
-
-	var response any
-	rh := func(r *http.Response) error {
-		if r.ContentLength == 0 {
-			return nil
-		}
-		if r.StatusCode == http.StatusNoContent {
-			return nil
-		}
-		if r.Body == nil {
-			return &httplib.StatusError{StatusCode: 404}
-		}
-		return httplib.FromJSON(&response)(r)
-	}
-
-	err = httplib.Fire(cli, req, httplib.FireOptions{
-		Verbose:         u.Verbose,
-		ResponseHandler: rh,
-		AuthMethod:      u.Auth,
-		Validators: []httplib.HandleResponseFunc{
-			httplib.ErrorJSON(apiErr, validStatusCodes...),
-		},
-	})
-	if err != nil {
-		return nil, err
-	}
-	val, ok = response.(map[string]interface{})
-	if !ok {
-		return nil, nil
-	}
-	return &val, nil
-}
-
-func (u *UnstructuredClient) Post(ctx context.Context, cli *http.Client, path string, opts *RequestConfiguration) (*map[string]interface{}, error) {
+func (u *UnstructuredClient) Call(ctx context.Context, cli *http.Client, path string, opts *RequestConfiguration) (any, error) {
 	uri := buildPath(u.Server, path, opts.Parameters, opts.Query)
 	pathItem, ok := u.DocScheme.Model.Paths.PathItems.Get(path)
 	if !ok {
 		return nil, fmt.Errorf("path not found: %s", path)
 	}
-	if len(pathItem.Post.Servers) > 0 {
-		uri = buildPath(pathItem.Post.Servers[0].URL, path, opts.Parameters, opts.Query)
-	}
+	httpMethod := string(opts.Method)
 
-	err := u.ValidateRequest("POST", path, opts.Parameters, opts.Query)
-	if err != nil {
-		return nil, err
-	}
-
-	req, err := httplib.Post(uri.String(), httplib.ToJSON(opts.Body))
-	if err != nil {
-		return nil, err
-	}
-	req.Header.Add("Content-Type", "application/json")
-
-	var val map[string]interface{}
-	apiErr := &APIError{}
-
-	httpMethod := "POST"
-
-	getDoc, ok := pathItem.GetOperations().Get(strings.ToLower(httpMethod))
-	if !ok {
-		return nil, fmt.Errorf("operation not found: %s", httpMethod)
-	}
-
-	validStatusCodes, err := getValidResponseCode(getDoc.Responses.Codes)
+	err := u.ValidateRequest(httpMethod, path, opts.Parameters, opts.Query)
 	if err != nil {
 		return nil, err
 	}
 
 	var response any
-	rh := func(r *http.Response) error {
-		if r.ContentLength == 0 {
-			return nil
+
+	var payload []byte
+
+	headers := make(http.Header)
+	payload = nil
+	m, ok := opts.Body.(map[string]any)
+	if !ok && opts.Body != nil {
+		return nil, fmt.Errorf("invalid body type: %T", opts.Body)
+	}
+	if len(m) != 0 {
+		jsonBody, err := json.Marshal(opts.Body)
+		if err != nil {
+			return nil, err
 		}
-		return httplib.FromJSON(&response)(r)
+		payload = jsonBody
+		headers.Set("Content-Type", "application/json")
 	}
 
-	err = httplib.Fire(cli, req, httplib.FireOptions{
-		Verbose:         u.Verbose,
-		ResponseHandler: rh,
-		AuthMethod:      u.Auth,
-		Validators: []httplib.HandleResponseFunc{
-			httplib.ErrorJSON(apiErr, validStatusCodes...),
-		},
-	})
+	req := &http.Request{
+		Method: httpMethod,
+		URL:    uri,
+		Proto:  "HTTP/1.1",
+		Body:   io.NopCloser(bytes.NewReader(payload)),
+		Header: headers,
+	}
+
+	if u.Debug {
+		cli.Transport = &debuggingRoundTripper{
+			Transport: cli.Transport,
+			Out:       os.Stdout,
+		}
+	}
+
+	if u.SetAuth != nil {
+		u.SetAuth(req)
+	}
+
+	resp, err := cli.Do(req)
+	if err != nil {
+		return nil, fmt.Errorf("error making request: %w", err)
+	}
+	defer resp.Body.Close()
+
+	getDoc, ok := pathItem.GetOperations().Get(strings.ToLower(httpMethod))
+	if !ok {
+		return nil, fmt.Errorf("operation not found: %s", httpMethod)
+	}
+	validStatusCodes, err := getValidResponseCode(getDoc.Responses.Codes)
 	if err != nil {
 		return nil, err
 	}
-	val, ok = response.(map[string]interface{})
+
+	if !HasValidStatusCode(resp.StatusCode, validStatusCodes...) {
+		return nil, &StatusError{
+			StatusCode: resp.StatusCode,
+			Inner:      fmt.Errorf("invalid status code: %d", resp.StatusCode),
+		}
+	}
+
+	err = handleResponse(resp.Body, &response)
+	if err != nil {
+		return nil, fmt.Errorf("error handling response: %w", err)
+	}
+
+	val, ok := response.(map[string]interface{})
 	if !ok {
 		return nil, nil
 	}
 	return &val, nil
 }
 
-func (u *UnstructuredClient) List(ctx context.Context, cli *http.Client, path string, opts *RequestConfiguration) (*map[string]interface{}, error) {
-	uri := buildPath(u.Server, path, opts.Parameters, opts.Query)
-	pathItem, ok := u.DocScheme.Model.Paths.PathItems.Get(path)
-	if !ok {
-		return nil, fmt.Errorf("path not found - list: %s", path)
-	}
-	if len(pathItem.Get.Servers) > 0 {
-		uri = buildPath(pathItem.Get.Servers[0].URL, path, opts.Parameters, opts.Query)
-	}
-
-	err := u.ValidateRequest("GET", path, opts.Parameters, opts.Query)
+func (u *UnstructuredClient) FindBy(ctx context.Context, cli *http.Client, path string, opts *RequestConfiguration) (any, error) {
+	list, err := u.Call(ctx, cli, path, opts)
 	if err != nil {
 		return nil, err
 	}
-	req, err := httplib.Get(uri.String())
-	if err != nil {
-		return nil, err
-	}
-
-	var val map[string]interface{}
-	apiErr := &APIError{}
-
-	httpMethod := "GET"
-	getDoc, ok := pathItem.GetOperations().Get(strings.ToLower(httpMethod))
-	if !ok {
-		return nil, fmt.Errorf("operation not found: %s", httpMethod)
-	}
-
-	validStatusCodes, err := getValidResponseCode(getDoc.Responses.Codes)
-	if err != nil {
-		return nil, err
-	}
-
-	var response any
-	rh := func(r *http.Response) error {
-		if r.ContentLength == 0 {
-			return nil
-		}
-		return httplib.FromJSON(&response)(r)
-	}
-
-	err = httplib.Fire(cli, req, httplib.FireOptions{
-		Verbose:         u.Verbose,
-		ResponseHandler: rh,
-		AuthMethod:      u.Auth,
-		Validators: []httplib.HandleResponseFunc{
-			httplib.ErrorJSON(apiErr, validStatusCodes...),
-		},
-	})
-	if err != nil {
-		return nil, err
-	}
-	val, ok = response.(map[string]interface{})
-	if !ok {
+	if list == nil {
 		return nil, nil
 	}
-	return &val, nil
-}
 
-func (u *UnstructuredClient) FindBy(ctx context.Context, cli *http.Client, path string, opts *RequestConfiguration) (*map[string]interface{}, error) {
-	list, err := u.List(ctx, cli, path, opts)
-	if err != nil {
-		return nil, err
+	var li map[string]interface{}
+
+	if _, ok := list.([]interface{}); !ok {
+		li = map[string]interface{}{
+			"items": list,
+		}
 	}
-	for _, v := range *list {
+
+	if _, ok := list.(map[string]interface{}); !ok {
+		return nil, fmt.Errorf("unexpected response type: %T", list)
+	}
+
+	for _, v := range li {
 		if v, ok := v.([]interface{}); ok {
 			if len(v) > 0 {
 				for _, item := range v {
 					if item, ok := item.(map[string]interface{}); ok {
-
 						for _, ide := range u.IdentifierFields {
 							idepath := strings.Split(ide, ".") // split the identifier field by '.'
 							responseValue, _, err := unstructured.NestedString(item, idepath...)
@@ -289,201 +149,58 @@ func (u *UnstructuredClient) FindBy(ctx context.Context, cli *http.Client, path 
 			break
 		}
 	}
-	return nil, &httplib.StatusError{StatusCode: 404}
+	return nil, &StatusError{
+		StatusCode: http.StatusNotFound,
+		Inner:      fmt.Errorf("item not found"),
+	}
 }
 
-func (u *UnstructuredClient) Patch(ctx context.Context, cli *http.Client, path string, opts *RequestConfiguration) (*map[string]interface{}, error) {
-	uri := buildPath(u.Server, path, opts.Parameters, opts.Query)
-	pathItem, ok := u.DocScheme.Model.Paths.PathItems.Get(path)
-	if !ok {
-		return nil, fmt.Errorf("path not found: %s", path)
+// buildPath constructs the URL path with the given parameters and query.
+// response should be a pointer to the object where the response will be unmarshalled.
+func handleResponse(rc io.ReadCloser, response any) error {
+	if rc == nil {
+		return nil
 	}
-	if len(pathItem.Patch.Servers) > 0 {
-		uri = buildPath(pathItem.Patch.Servers[0].URL, path, opts.Parameters, opts.Query)
-	}
-
-	err := u.ValidateRequest("PATCH", path, opts.Parameters, opts.Query)
+	defer rc.Close()
+	data, err := io.ReadAll(rc)
 	if err != nil {
-		return nil, err
+		return err
 	}
-
-	req, err := httplib.Patch(uri.String(), httplib.ToJSON(opts.Body))
-	if err != nil {
-		return nil, err
+	if len(data) == 0 {
+		return nil
 	}
-	req.Header.Add("Content-Type", "application/json")
-
-	var val map[string]interface{}
-	apiErr := &APIError{}
-
-	httpMethod := "PATCH"
-
-	getDoc, ok := pathItem.GetOperations().Get(strings.ToLower(httpMethod))
-	if !ok {
-		return nil, fmt.Errorf("operation not found: %s", httpMethod)
-	}
-
-	validStatusCodes, err := getValidResponseCode(getDoc.Responses.Codes)
-	if err != nil {
-		return nil, err
-	}
-
-	var response any
-	rh := func(r *http.Response) error {
-		if r.ContentLength == 0 {
-			return nil
-		}
-		return httplib.FromJSON(&response)(r)
-	}
-
-	err = httplib.Fire(cli, req, httplib.FireOptions{
-		Verbose:         u.Verbose,
-		ResponseHandler: rh,
-		AuthMethod:      u.Auth,
-		Validators: []httplib.HandleResponseFunc{
-			httplib.ErrorJSON(apiErr, validStatusCodes...),
-		},
-	})
-	if err != nil {
-		return nil, err
-	}
-	val, ok = response.(map[string]interface{})
-	if !ok {
-		return nil, nil
-	}
-	return &val, nil
+	return json.Unmarshal(data, &response)
 }
 
-func (u *UnstructuredClient) Put(ctx context.Context, cli *http.Client, path string, opts *RequestConfiguration) (*map[string]interface{}, error) {
-	uri := buildPath(u.Server, path, opts.Parameters, opts.Query)
-	pathItem, ok := u.DocScheme.Model.Paths.PathItems.Get(path)
-	if !ok {
-		return nil, fmt.Errorf("path not found: %s", path)
-	}
-	if len(pathItem.Put.Servers) > 0 {
-		uri = buildPath(pathItem.Put.Servers[0].URL, path, opts.Parameters, opts.Query)
-	}
-
-	err := u.ValidateRequest("PUT", path, opts.Parameters, opts.Query)
-	if err != nil {
-		return nil, err
-	}
-
-	req, err := httplib.Put(uri.String(), httplib.ToJSON(opts.Body))
-	if err != nil {
-		return nil, err
-	}
-	req.Header.Add("Content-Type", "application/json")
-
-	var val map[string]interface{}
-	apiErr := &APIError{}
-
-	httpMethod := "PUT"
-	getDoc, ok := pathItem.GetOperations().Get(strings.ToLower(httpMethod))
-	if !ok {
-		return nil, fmt.Errorf("operation not found: %s", httpMethod)
-	}
-
-	validStatusCodes, err := getValidResponseCode(getDoc.Responses.Codes)
-	if err != nil {
-		return nil, err
-	}
-
-	var response any
-	rh := func(r *http.Response) error {
-		if r.ContentLength == 0 {
-			return nil
-		}
-		return httplib.FromJSON(&response)(r)
-	}
-
-	if containsStatusCode(http.StatusNoContent, validStatusCodes) {
-		rh = nil
-	}
-
-	err = httplib.Fire(cli, req, httplib.FireOptions{
-		Verbose:         u.Verbose,
-		ResponseHandler: rh,
-		AuthMethod:      u.Auth,
-		Validators: []httplib.HandleResponseFunc{
-			httplib.ErrorJSON(apiErr, validStatusCodes...),
-		},
-	})
-	if err != nil {
-		return nil, err
-	}
-	val, ok = response.(map[string]interface{})
-	if !ok {
-		return nil, nil
-	}
-	return &val, nil
+type debuggingRoundTripper struct {
+	Transport http.RoundTripper
+	Out       io.Writer
 }
 
-func (u *UnstructuredClient) Delete(ctx context.Context, cli *http.Client, path string, opts *RequestConfiguration) (*map[string]interface{}, error) {
-	uri := buildPath(u.Server, path, opts.Parameters, opts.Query)
-	pathItem, ok := u.DocScheme.Model.Paths.PathItems.Get(path)
-	if !ok {
-		return nil, fmt.Errorf("path not found: %s", path)
-	}
-	if len(pathItem.Delete.Servers) > 0 {
-		uri = buildPath(pathItem.Delete.Servers[0].URL, path, opts.Parameters, opts.Query)
-	}
-
-	err := u.ValidateRequest("DELETE", path, opts.Parameters, opts.Query)
-	if err != nil {
-		return nil, err
-	}
-	req, err := httplib.Delete(uri.String())
+func (d *debuggingRoundTripper) RoundTrip(req *http.Request) (*http.Response, error) {
+	b, err := httputil.DumpRequestOut(req, true)
 	if err != nil {
 		return nil, err
 	}
 
-	var val map[string]interface{}
-	apiErr := &APIError{}
+	d.Out.Write(b)
+	d.Out.Write([]byte{'\n'})
 
-	httpMethod := "DELETE"
-	getDoc, ok := pathItem.GetOperations().Get(strings.ToLower(httpMethod))
-	if !ok {
-		return nil, fmt.Errorf("operation not found: %s", httpMethod)
+	if d.Transport == nil {
+		d.Transport = http.DefaultTransport
 	}
 
-	validStatusCodes, err := getValidResponseCode(getDoc.Responses.Codes)
+	resp, err := d.Transport.RoundTrip(req)
+	if err != nil {
+		return resp, err
+	}
+
+	b, err = httputil.DumpResponse(resp, req.URL.Query().Get("watch") != "true")
 	if err != nil {
 		return nil, err
 	}
+	d.Out.Write(b)
+	d.Out.Write([]byte{'\n'})
 
-	var response any
-	rh := func(r *http.Response) error {
-		if r.ContentLength == 0 {
-			return nil
-		}
-		return httplib.FromJSON(&response)(r)
-	}
-
-	err = httplib.Fire(cli, req, httplib.FireOptions{
-		Verbose:         u.Verbose,
-		ResponseHandler: rh,
-		AuthMethod:      u.Auth,
-		Validators: []httplib.HandleResponseFunc{
-			httplib.ErrorJSON(apiErr, validStatusCodes...),
-		},
-	})
-	if err != nil {
-		return nil, err
-	}
-
-	val, ok = response.(map[string]interface{})
-	if !ok {
-		return nil, nil
-	}
-	return &val, nil
-}
-
-func containsStatusCode(statusCode int, validStatusCodes []int) bool {
-	for _, v := range validStatusCodes {
-		if v == statusCode {
-			return true
-		}
-	}
-	return false
+	return resp, err
 }
